AS726X and other Ghana improvements

+++ LOOM_library/loom_analogIn.h
@@ -8,7 +8,8 @@
 // ===                        STRUCTURES                        === 
 // ================================================================
 struct state_analog_t {
-	int a0, a1, a2, a3, a4, a5; // Memory to store analog sensor values
+	// floats to support analog conversion functions
+	float a0, a1, a2, a3, a4, a5; // Memory to store analog sensor values
 													// Add other ports ? put these in #defines based on which ports are enabled
 };
 
@@ -165,29 +166,53 @@ void package_analog(OSCBundle *bndl, char packet_header_string[])
 	char address_string[80];
 
 	// Get readings from enabled ports           
+	// #if (is_analog_a0 == 1) 
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A0");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a0);
+	// #endif
+	// #if (is_analog_a1 == 1)
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A1");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a1);
+	// #endif
+	// #if (is_analog_a2 == 1) 
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A2");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a2);
+	// #endif
+	// #if (is_analog_a3 == 1)
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A3");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a3);
+	// #endif
+	// #if (is_analog_a4 == 1) 
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A4");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a4);
+	// #endif
+	// #if (is_analog_a5 == 1)
+	// 	sprintf(address_string, "%s%s", packet_header_string, "/pin_A5");
+	// 	bndl->add(address_string).add((int32_t)state_analog.a5);
+	// #endif
 	#if (is_analog_a0 == 1) 
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A0");
-		bndl->add(address_string).add((int32_t)state_analog.a0);
+		bndl->add(address_string).add(state_analog.a0);
 	#endif
 	#if (is_analog_a1 == 1)
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A1");
-		bndl->add(address_string).add((int32_t)state_analog.a1);
+		bndl->add(address_string).add(state_analog.a1);
 	#endif
 	#if (is_analog_a2 == 1) 
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A2");
-		bndl->add(address_string).add((int32_t)state_analog.a2);
+		bndl->add(address_string).add(state_analog.a2);
 	#endif
 	#if (is_analog_a3 == 1)
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A3");
-		bndl->add(address_string).add((int32_t)state_analog.a3);
+		bndl->add(address_string).add(state_analog.a3);
 	#endif
 	#if (is_analog_a4 == 1) 
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A4");
-		bndl->add(address_string).add((int32_t)state_analog.a4);
+		bndl->add(address_string).add(state_analog.a4);
 	#endif
 	#if (is_analog_a5 == 1)
 		sprintf(address_string, "%s%s", packet_header_string, "/pin_A5");
-		bndl->add(address_string).add((int32_t)state_analog.a5);
+		bndl->add(address_string).add(state_analog.a5);
 	#endif
 }
